requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
cloudinary==1.39.0
Django==4.2.4
django-cloudinary-storage==0.3.0
django-cors-headers==4.2.0
djangorestframework==3.14.0
gunicorn==21.2.0
idna==3.6
packaging==23.1
pillow==10.2.0
psycopg2==2.9.6
python-dotenv==1.0.1
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.2.1
whitenoise==6.5.0
